add track url when get playlist
